libraries upgrade and cleanup

+++ src/config.cpp
@@ -1,7 +1,7 @@
 #include "config.h"
 namespace art
 {
-    #include "board.h"
+#include "hw/board.h"
     Config::Config()
     {
         if (ESP_FS.exists(CONFIG_FILE))
@@ -77,7 +77,7 @@ namespace art
 
     void Config::cleanupWiFi()
     {
-        for (size_t i = 0; i < wifi.size(); i++)
+        for (int i = 0; i < wifi.size(); i++)
         {
             delete wifi.get(i);
         }
@@ -86,13 +86,12 @@ namespace art
 
     void Config::cleanupDmx()
     {
-        for (size_t i = 0; i < dmx.size(); i++)
+        for (int i = 0; i < dmx.size(); i++)
         {
             delete dmx.get(i);
         }
         dmx.clear();
     }
-
 
     void Config::cleanup()
     {
@@ -183,9 +182,9 @@ namespace art
         doc["info"]["rssi"] = WiFi.RSSI();
         doc["id"] = CHIP_ID;
         doc["host"] = host;
-        
+
         JsonArray wifiNets = doc.createNestedArray("wifi");
-        for (size_t i = 0; i < wifi.size(); i++)
+        for (int i = 0; i < wifi.size(); i++)
         {
             WiFiNet *net = wifi.get(i);
             wifiNets[i]["ssid"] = net->ssid;
@@ -194,7 +193,7 @@ namespace art
             wifiNets[i]["order"] = net->order;
         }
         JsonArray channels = doc.createNestedArray("dmx");
-        for (size_t i = 0; i < dmx.size() && i < MAX_DMX_DEVICES; i++)
+        for (int i = 0; i < dmx.size() && i < MAX_DMX_DEVICES; i++)
         {
             DmxChannel *channel = dmx.get(i);
             channels[i]["channel"] = channel->channel;
